+++ core/src/main/java/arez/annotations/ArezComponent.java
@@ -186,4 +186,13 @@ public @interface ArezComponent
    * @return enum controlling whether a references to components should be explicitly managed.
    */
   Feature verifyReferencesToComponent() default Feature.AUTODETECT;
+
+  /**
+   * The default priority used by {@link Memoize} and {@link Observe} annotated methods.
+   * This parameter should not be specified only be specified
+   * if there are {@link Memoize} or {@link Observe} annotated methods present on the component.
+   *
+   * @return the default priority used by {@link Memoize} and {@link Observe} annotated methods.
+   */
+  Priority defaultPriority() default Priority.NORMAL;
 }