+++ processor/src/main/java/arez/processor/Priority.java
@@ -1,0 +1,10 @@
+package arez.processor;
+
+enum Priority
+{
+  HIGHEST,
+  HIGH,
+  NORMAL,
+  LOW,
+  LOWEST
+}