+++ processor/src/test/fixtures/input/com/example/component/Suppressed1UnnecessaryDefaultPriorityPresentComponent.java
@@ -1,0 +1,11 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Priority;
+
+// This uses the SOURCE retention suppression
+@SuppressWarnings( "Arez:UnnecessaryDefaultPriority" )
+@ArezComponent( allowEmpty = true, defaultPriority = Priority.LOW )
+public abstract class Suppressed1UnnecessaryDefaultPriorityPresentComponent
+{
+}