+++ processor/src/test/fixtures/input/com/example/component/Suppressed2UnnecessaryDefaultPriorityPresentComponent.java
@@ -1,0 +1,12 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Priority;
+import arez.annotations.SuppressArezWarnings;
+
+// This uses the CLASS retention suppression
+@SuppressArezWarnings( "Arez:UnnecessaryDefaultPriority" )
+@ArezComponent( allowEmpty = true, defaultPriority = Priority.LOW )
+public abstract class Suppressed2UnnecessaryDefaultPriorityPresentComponent
+{
+}