+++ processor/src/test/fixtures/input/com/example/component/UnnecessaryDefaultPriorityPresentComponent.java
@@ -1,0 +1,9 @@
+package com.example.component;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Priority;
+
+@ArezComponent( allowEmpty = true, defaultPriority = Priority.LOW )
+public abstract class UnnecessaryDefaultPriorityPresentComponent
+{
+}