+++ processor/src/test/fixtures/input/com/example/memoize/DefaultPriorityUnspecifiedLocalPriorityMemoizeModel.java
@@ -1,0 +1,15 @@
+package com.example.memoize;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Memoize;
+import arez.annotations.Priority;
+
+@ArezComponent( defaultPriority = Priority.HIGH )
+abstract class DefaultPriorityUnspecifiedLocalPriorityMemoizeModel
+{
+  @Memoize
+  public long getTime()
+  {
+    return 0;
+  }
+}