+++ processor/src/test/fixtures/input/com/example/observe/DefaultPriorityUnspecifiedLocalPriorityObserveModel.java
@@ -1,0 +1,14 @@
+package com.example.observe;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Observe;
+import arez.annotations.Priority;
+
+@ArezComponent( defaultPriority = Priority.HIGH )
+abstract class DefaultPriorityUnspecifiedLocalPriorityObserveModel
+{
+  @Observe
+  void doStuff()
+  {
+  }
+}